requirements.txt
Django==4.2.7
gunicorn==21.2.0
psycopg2-binary==2.9.7
pillow==10.4.0
dj-database-url==2.0.0
python-dotenv==1.0.0
